+++ src/app/components/favorites/favorites.tsx
@@ -1,7 +1,7 @@
 import { useSelector } from "react-redux";
 import { RootState } from "@/app/stores";
 import EmptyFavorites from "./emptyFavorites";
-import FavoritesContent from "./FavoritesContent";
+import FavoritesContent from "./favoritesContent";
 
 export default function Favorites() {
   const favorites = useSelector((state: RootState) => state.favorites);
